Server/requirements.txt
fastapi
langchain
langchain-core
langchain-groq
langserve
pydantic
python-dotenv
uvicorn